Refactor: request, response 패키지별 분리 (#23)
