+++ src/main/java/com/kgu/life_watch/domain/auth/controller/AuthSmsController.java
@@ -6,7 +6,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 
-import com.kgu.life_watch.domain.auth.dto.SmsVerificationRequest;
+import com.kgu.life_watch.domain.auth.dto.request.SmsVerificationRequest;
+import com.kgu.life_watch.domain.auth.dto.response.SmsVerifyResponse;
 import com.kgu.life_watch.domain.auth.service.AuthSmsService;
 import com.kgu.life_watch.global.domain.SuccessCode;
 import com.kgu.life_watch.global.dto.response.ApiResponse;
@@ -22,7 +23,7 @@ public class AuthSmsController {
   // 인증 메세지 발송
   @GetMapping("/sms")
   @Operation(summary = "인증 메시지 발송 API", description = "인증 메시지를 발송하는 API입니다.")
-  public ApiResponse<String> sendSms(@RequestParam String phone) {
+  public ApiResponse<Void> sendSms(@RequestParam String phone) {
     smsService.sendAuthenticationCode(phone);
     return new ApiResponse<>(SuccessCode.REQUEST_OK);
   }
@@ -30,8 +31,8 @@ public class AuthSmsController {
   // 인증
   @PostMapping("/sms/verify")
   @Operation(summary = "인증 메시지 검증 API", description = "인증 메시지를 검증하는 API입니다.")
-  public ApiResponse<Boolean> verifyCode(@RequestBody SmsVerificationRequest request) {
+  public ApiResponse<SmsVerifyResponse> verifyCode(@RequestBody SmsVerificationRequest request) {
     boolean isValid = smsService.verifyCode(request.phoneNumber(), request.verificationCode());
-    return new ApiResponse<>(isValid);
+    return new ApiResponse<>(SmsVerifyResponse.builder().isValid(isValid).build());
   }
 }
