+++ src/main/java/com/kgu/life_watch/domain/auth/dto/request/FcmTokenUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.kgu.life_watch.domain.auth.dto;
+package com.kgu.life_watch.domain.auth.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
