+++ src/main/java/com/kgu/life_watch/domain/auth/dto/request/FindIdRequest.java
@@ -1,3 +1,3 @@
-package com.kgu.life_watch.domain.auth.dto;
+package com.kgu.life_watch.domain.auth.dto.request;
 
 public record FindIdRequest(String name, String phoneNumber) {}
