+++ src/main/java/com/kgu/life_watch/domain/auth/dto/request/ResetPasswordRequest.java
@@ -1,3 +1,3 @@
-package com.kgu.life_watch.domain.auth.dto;
+package com.kgu.life_watch.domain.auth.dto.request;
 
 public record ResetPasswordRequest(String loginId, String newPassword, String verificationCode) {}
