+++ src/main/java/com/kgu/life_watch/domain/auth/dto/response/FindLoginIdResponse.java
@@ -1,4 +1,4 @@
-package com.kgu.life_watch.domain.auth.dto;
+package com.kgu.life_watch.domain.auth.dto.response;
 
 import lombok.Builder;
