+++ src/main/java/com/kgu/life_watch/domain/auth/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.kgu.life_watch.domain.auth.dto;
+package com.kgu.life_watch.domain.auth.dto.response;
 
 import java.time.LocalDate;
