+++ src/main/java/com/kgu/life_watch/domain/auth/dto/response/SmsVerifyResponse.java
@@ -1,0 +1,6 @@
+package com.kgu.life_watch.domain.auth.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record SmsVerifyResponse(Boolean isValid) {}
